-Added quit game warning message. -Updated requirements

+++ src/ui/offlineNewGame.py
@@ -1,4 +1,5 @@
 import customtkinter as ctk
+from CTkMessagebox import CTkMessagebox
 from chess.Chess import Chess
 
 class OfflineNewGame(ctk.CTkFrame):
@@ -9,24 +10,34 @@ class OfflineNewGame(ctk.CTkFrame):
         self.grid_rowconfigure(0, weight=1)
         self.grid_columnconfigure(0, weight=1)
 
-        self.side_bar = ctk.CTkFrame(self)
+        self.side_bar = ctk.CTkTabview(self)
+        self.side_bar.add("Game")
+        self.side_bar.add("History")
         self.side_bar.grid(row=0, column=1, sticky="news")
 
-        self.quit_button = ctk.CTkButton(self.side_bar, text="Quit", command=self.quitFrame)
-        self.quit_button.grid(row=0, column=0)
+        self.quit_button = ctk.CTkButton(self.side_bar.tab("Game"), text="Quit", command=self.quitFrame)
+        self.quit_button.grid(row=0, column=0, padx=(10, 0))
+
+        self.refresh_button = ctk.CTkButton(self.side_bar.tab("Game"), text="Refresh", command=self.updateBoard)
+        self.refresh_button.grid(row=0, column=1, padx=10)
 
         self.board_frame = ctk.CTkFrame(self, fg_color="transparent")
         self.board_frame.grid(row=0, column=0, sticky="nesw")
-        self.board_frame.bind("<Configure>", self.updateBoard)
+        self.board_frame.bind("<Configure>", lambda e: self.updateBoard())
 
         self.chess = Chess(self.board_frame)
     
     def quitFrame(self): #add confirmation box
-        self.root.setActivePage("Home", "Home")
-        self.root.deleteFrame("OfflineNewGame")
-        self.root.showFrame("Home")
+        warning_popup = CTkMessagebox(title="Quit", message="Do you want to quit?",
+                        icon="question", option_1="No", option_2="Yes")
 
-    def updateBoard(self, e):
+        if warning_popup.get() == "Yes":
+            self.root.setActivePage("Home", "Home")
+            self.root.deleteFrame("OfflineNewGame")
+            self.root.showFrame("Home")
+
+    def updateBoard(self):
         self.root.update()
         self.chess.updateGame()
+
